agentic_hate_rag/requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.1
requests>=2.31.0
langchain>=0.2.16
langchain-core>=0.2.38
langchain-community>=0.2.16
langchain-chroma>=0.1.4
langchain-ollama>=0.1.3
langgraph>=0.2.19
chromadb>=0.5.5
transformers>=4.44.0
torch>=2.2.0
numpy>=1.26.0
indic-transliteration>=2.3.60
